+++ src/lib/services/block/jsonRpc.ts
@@ -1,9 +1,15 @@
-import { requestJsonRpc } from "../evm/jsonRpc";
-import { zBlockJsonRpc } from "../types";
+import { requestBatchJsonRpc } from "../evm/jsonRpc";
+import { zBlockDataJsonRpc } from "../types";
 import { numberToHex, parseWithError } from "lib/utils";
 
 export const getBlockDataJsonRpc = (endpoint: string, height: number) =>
-  requestJsonRpc(endpoint, "eth_getBlockByNumber", [
-    numberToHex(height),
-    true,
-  ]).then(({ result }) => parseWithError(zBlockJsonRpc, result));
+  requestBatchJsonRpc(endpoint, [
+    {
+      method: "eth_getBlockByNumber",
+      params: [numberToHex(height), true],
+    },
+    {
+      method: "eth_getBlockReceipts",
+      params: [numberToHex(height)],
+    },
+  ]).then((results) => parseWithError(zBlockDataJsonRpc, results));