+++ src/lib/services/types/evm/index.ts
@@ -1,2 +1,2 @@
-export * from "./params";
 export * from "./jsonRpc";
+export * from "./params";